+++ config/routes.rb
@@ -1,7 +1,7 @@
 Rails.application.routes.draw do
   root "slides#new"
 
-  resources :slides, only: [ :new, :create, :show ]
+  resources :slides, only: [ :new, :create, :show, :update ]
 
   # Reveal health status on /up that returns 200 if the app boots with no exceptions, otherwise 500.
   # Can be used by load balancers and uptime monitors to verify that the app is live.